server/requirements.txt
dataframe_image==0.1.10
Flask==2.2.2
Flask_MySQLdb==1.0.1
keras==2.10.0
moviepy==1.0.3
mysql-connector==2.2.9
mysql-connector-python==8.0.31
mysqlclient==2.2.1
mysql_connector_repackaged==0.3.1
numpy==1.24.0
opencv_python==4.6.0.66
pandas==1.5.2
python-dotenv==1.0.0
pytube==12.1.3
tensorflow_gpu==2.10.1
